UnionFindTree-iin size-iig ni urjihed integer overflow uusej bsan uchir ll bolgov

+++ contest/main.cpp
@@ -147,7 +147,7 @@ public:
 		return true;
 	}
 
-	int size(int a) {
+	ll size(int a) {
 		return union_size[root(a)];
 	}
